tests_ignore/test_plus_minus/plots/requirements.txt
21cmFAST @ file:///leonardo_work/CNHPC_1497299/ntriantafyllou/database/database3_venv/21cmFAST
astropy==6.1.2
astropy-iers-data==0.2024.8.19.0.32.16
astropy_healpix==1.0.3
attrs==24.2.0
bidict==0.23.1
cached-property==1.5.2
camb==1.5.8
cffi==1.17.0
click==8.1.7
contourpy==1.2.1
cosmotile==0.2.4
cycler==0.12.1
deprecation==2.1.0
fonttools==4.53.1
h5py==3.11.0
hmf==3.5.0
jsons==1.6.3
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
numpy==1.26.4
packaging==24.1
pillow==10.4.0
powerbox==0.8.1
psrecord==1.4
psutil==6.0.0
py21cmfast_tools==1.0.4
pycparser==2.22
pyerfa==2.0.1.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rich==13.8.1
scipy==1.14.1
six==1.16.0
sympy==1.13.3
toml==0.10.2
typing_extensions==4.12.2
typish==1.9.3
